+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 ### Added
+### Fixed
+
+## [5.6.0] - 2024-06-15
+### Added
 - [Issue 485](https://github.com/aza547/wow-recorder/issues/485) - Added upload rate limit setting.
 
 ### Fixed
-- Make the chat overlay scale slider step size smaller. 
+- Make the chat overlay scale slider step size smaller.
+- Fix the scrollbar width on the scene editor. 
 
 ## [5.5.0] - 2024-06-12
 ### Added